Add tests for the filesystem, fsstats, and process MetricSets. (#1566)
- Fix issue where memory user_p and swap_p were not being calculated.
- Fix error where proc.cpu.total_p value could be NaN if the polling interval is less that 0.5 ms. When less than 0.5 ms, the time delta value was truncated to 0. Then 0 was being used as the denominator in the percentage calculation causing it to result in NaN. I changed the calculation to use a float64 for the time delta instead of an int64 so that the value is never truncated to 0.
- Fix golint warnings in the system module.
- Add missing fields to fields.yml (filesystem was missing free, and fsstats was missing total_size.total).
- Change data type in fields.yml for several fields marked as integer that are uint64 or int64 in Go.

+++ topbeat/system/cpu.go
@@ -78,7 +78,7 @@ func GetCpuPercentageList(last, current []CpuTimes) []CpuTimes {
 		calculate := func(field2 uint64, field1 uint64, all_delta uint64) float64 {
 
 			perc := 0.0
-			delta := field2 - field1
+			delta := int64(field2 - field1)
 			perc = float64(delta) / float64(all_delta)
 			return Round(perc, .5, 4)
 		}
